推送 ------------- 推送日期 2021/12/20 ------------- 目前進度
分類列表頁功能完成
ADD、EDIT超連結可作
可刪除流水帳(有紀錄之分類不可刪除)
顯示流水帳
分頁功能

+++ Accounting-system/src/main/java/com/ubayKyu/accountingSystem/controller/CategoryListController.java
@@ -3,17 +3,18 @@ package com.ubayKyu.accountingSystem.controller;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.ubayKyu.accountingSystem.dto.CategoryInterface;
 import com.ubayKyu.accountingSystem.entity.CategoryModel;
 import com.ubayKyu.accountingSystem.repository.CategoryRepository;
 
@@ -21,6 +22,7 @@ import com.ubayKyu.accountingSystem.repository.CategoryRepository;
 public class CategoryListController {
 	@Autowired
 	CategoryRepository categoryRepository;
+
 
 	@GetMapping("/CategoryList")
 	public String CategoryList(Model model, HttpServletRequest request) {
@@ -33,19 +35,21 @@ public class CategoryListController {
 			// 取得登入使用者的ID去DB內尋找該使用者的自定義分類
 			String userid = request.getSession().getAttribute("userid").toString();
 
-			List<CategoryModel> list = categoryRepository.getFindAll(userid);
-			model.addAttribute("list", list);
+			List<CategoryInterface> list = categoryRepository.getFindAll(userid);
+			model.addAttribute("list", list);			
 			return "CategoryList";
 		}
 
 	}
-	@PostMapping
+
+
 	@RequestMapping(value = "/CategoryList", method = RequestMethod.POST)
 	public String CategoryListDel(@RequestParam(value = "cblDel", required = false) String[] categoryDel,
-			@RequestParam(value = "btnAdd", required = false) String btnAdd,
-			HttpServletRequest request) {
+			@RequestParam(value = "btnAdd", required = false) String btnAdd, HttpServletRequest request,
+			RedirectAttributes redirAttrs) {
+	
 		String temp = btnAdd;
-		if(temp != "") {
+		if (temp != null) {
 			return "CategoryDetail";
 		}
 		// 登入判斷
@@ -56,16 +60,25 @@ public class CategoryListController {
 		} else {
 			// 確認有無清單選取
 			if (categoryDel != null) {
-				for (String categoryid : categoryDel) //用for跑字串陣列內的數值
+				for (String categoryid : categoryDel) // 用for跑字串陣列內的數值
 				{
-					categoryRepository.deleteCat(categoryid);
+					// 檢查選中分類內有無紀錄
+					int count = categoryRepository.checkCateCount(categoryid);
+					if(count == 0) {
+						
+						categoryRepository.deleteCat(categoryid);
+						redirAttrs.addFlashAttribute("message", "刪除成功");
 					
+					}
+					else {
+						redirAttrs.addFlashAttribute("message", "刪除失敗");
+						
+					}
 				}
-			} else {
-
-			}
+			} 
 		}
-		return "CategoryList";
+		
+		return "redirect:/CategoryList";
 	}
 
 }
